+++ packages/asc-ui/src/index.ts
@@ -9,6 +9,7 @@ import Tag, { TagStyle } from './components/Tag'
 import Link, { LinkStyle } from './components/Link'
 import Paragraph, { ParagraphStyle } from './components/Paragraph'
 import TopBar, { TopBarStyles } from './components/TopBar'
+import TopTaskButton, { TopTaskButtonStyle } from './components/TopTaskButton'
 import ShareButton from './components/Button/ShareButton'
 import { Article, ArticleStyle } from './components/Editorial/Article'
 import Button from './components/Button'
@@ -156,6 +157,7 @@ export const styles = {
   CustomHTMLBlockStyle,
   ArticleStyle,
   TagStyle,
+  TopTaskButtonStyle,
   CheckboxStyle,
   FilterBoxStyle,
   FilterOptionStyle,
@@ -242,6 +244,7 @@ export {
   ThemeProvider,
   CompactThemeProvider,
   TopBar,
+  TopTaskButton,
   EditorialPost,
   Article,
   Link,