+++ CHANGELOG.md
@@ -16,14 +16,13 @@ Prefix the change with one of these keywords:
 
 ## Unreleased
 - Changed: Internal SVG icon imports are replaced with React components, eliminating the need for including an SVG loader (closes #459)
-
+- Added: Added TopTaskButton component
 - Added: `theme` prop to `ThemeProvider` to pass a custom theme
 - Changed: **BREAKING** `createTheme` method from `Theme`.
 - Changed: **BREAKING** removed `themeName` from `ThemeProvider`
 - Changed: **BREAKING** removed `asc-core` package. `Theme` and `ascDefaultTheme` can now be found in `asc-ui`
 
 ## [0.19.2-beta.1]
-
 - Added Hidden component
 
 ## [0.19.1]